OPA rule to restrict updates of droplets to labels and annotations
The provided rule restricts the space-developer role members to only
be able to update labels and annotations in the metadata

The space developers role members are dteremined as follows:
- there is a role binding binding the user to the space developers role
- there is a role binding binding a group that the user belongs to to
  the space developers role

The group information is coming from a modified version of dex that puts
it as a claim in the issued token. K8s infers all user and group
information from the trusted token.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>